# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/user": {
      "get": {
        "operationId": "UserController_getUserProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User info recived",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnUserProfileDetails"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error"
          },
          "4XX": {
            "description": "Check response message"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user info",
        "tags": [
          "User"
        ]
      },
      "patch": {
        "operationId": "UserController_updateUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User info updated"
          },
          "5XX": {
            "description": "Server error"
          },
          "4XX": {
            "description": "Check response message"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user info",
        "tags": [
          "User"
        ]
      }
    },
    "/user/password": {
      "patch": {
        "operationId": "UserController_updateUserPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User password updated"
          },
          "5XX": {
            "description": "Server error"
          },
          "4XX": {
            "description": "Check response message"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user password",
        "tags": [
          "User"
        ]
      }
    },
    "/user/cart": {
      "get": {
        "operationId": "CartController_getCart",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items limit on page ( from 1 to 50 )",
            "schema": {
              "default": 10,
              "example": 15,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Items page ( min 1 )",
            "schema": {
              "default": 1,
              "example": 2,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User cart recived",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartDetails"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error"
          },
          "4XX": {
            "description": "Check response message"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user cart",
        "tags": [
          "Cart"
        ]
      },
      "post": {
        "operationId": "CartController_addProductToUserCart",
        "parameters": [
          {
            "name": "pid",
            "required": true,
            "in": "query",
            "description": "Product id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success added"
          },
          "5XX": {
            "description": "Server error"
          },
          "4XX": {
            "description": "Check response message"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add product to user cart",
        "tags": [
          "Cart"
        ]
      },
      "delete": {
        "operationId": "CartController_removeProductFromCart",
        "parameters": [
          {
            "name": "pid",
            "required": true,
            "in": "query",
            "description": "Product id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success removed"
          },
          "5XX": {
            "description": "Server error"
          },
          "4XX": {
            "description": "Check response message"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove product from user cart",
        "tags": [
          "Cart"
        ]
      }
    },
...
  },
  "info": {
    "title": "Fake Store API",
    "description": "A simple REST API that simulates an online store — provides fake products, categories, and users for testing, prototyping, and learning purposes.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ReturnUserProfileDetails": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "description": "User id"
          },
          "firstName": {
            "type": "string",
            "description": "User first name"
          },
          "lastName": {
            "type": "string",
            "description": "User last name"
          },
          "login": {
            "type": "string",
            "description": "User login"
          },
          "email": {
            "type": "string",
            "description": "User email"
          },
          "avatarURL": {
            "type": "string",
            "description": "User avatar url"
          }
        },
        "required": [
          "uuid",
          "firstName",
          "lastName",
          "login",
          "email",
          "avatarURL"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "User first name ( from 1 to 20 symbols )",
            "example": "Evgeny"
          },
          "lastName": {
            "type": "string",
            "description": "User last name ( from 1 to 20 symbols )",
            "example": "Smirnov"
          },
          "login": {
            "type": "string",
            "description": "User login ( from 3 to 24 symbols )",
            "example": "nodewalker"
          },
          "email": {
            "type": "string",
            "description": "User email",
            "example": "nodewalker@yandex.com"
          },
          "avatar": {
            "type": "string",
            "format": "binary",
            "description": "User avatar"
          }
        }
      },
      "UpdateUserPasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "description": "User old password ( from 8 to 32 symbols )",
            "example": "**********"
          },
          "newPassword": {
            "type": "string",
            "description": "User new password ( from 8 to 32 symbols )",
            "example": "**********"
          },
          "repeatNewPassword": {
            "type": "string",
            "description": "Repeat new password ( from 8 to 32 symbols )",
            "example": "**********"
          }
        },
        "required": [
          "currentPassword",
          "newPassword",
          "repeatNewPassword"
        ]
      },
      "ProductImageEntity": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "description": "Image url"
          }
        },
        "required": [
          "uuid"
        ]
      },
      "CategoryDetails": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "description": "Category id"
          },
          "name": {
            "type": "string",
            "description": "Category name"
          },
          "hasChildren": {
            "type": "boolean",
            "description": "Is category has children"
          },
          "hasProduct": {
            "type": "boolean",
            "description": "Is category has product"
          }
        },
        "required": [
          "uuid",
          "name"
        ]
      },
      "ProductDetails": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "description": "Product id"
          },
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "price": {
            "type": "string",
            "description": "Product price"
          },
          "discount": {
            "type": "string",
            "description": "Product discount"
          },
          "images": {
            "description": "Product images",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductImageEntity"
            }
          },
          "category": {
            "description": "Product category",
            "allOf": [
              {
                "$ref": "#/components/schemas/CategoryDetails"
              }
            ]
          }
        },
        "required": [
          "uuid",
          "name",
          "price",
          "discount",
          "images",
          "category"
        ]
      },
      "CartDetails": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "description": "Cart id"
          },
          "products": {
            "description": "Products from cart",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDetails"
            }
          }
        },
        "required": [
          "uuid",
          "products"
        ]
      },
...
    }
  }
}